[refactor&style] （还未完成）功能基本完成，能够下载下来文件，但是速度非常不理想，可能要开多任务。下载完成后的清理工作有 bug，没有正确释放资源，还需要排查。
下一步：
 1. 实现磁力链接下载
 2. [重要] - 观察内存使用情况
 3. 继续实现后续基本接口
 4. [重要] - 梳理 peer piece 请求和响应那块的逻辑，着重检查是否会出现分片偏移永久跳过的bug
 5. [重要] - 继续实现 task 以及 peer 部分的重构
 6. [重要] - 下载速度优化
 7. [重要] - 排查任务资源未正确释放的问题

+++ doro-util/src/bytes_util.rs
@@ -72,9 +72,9 @@ pub trait Bytes2Int<T> {
 impl_bytes2int!(u8, u16, u32, u64, u128);
 
 /// 1 字节位图数组偏移计算
-pub fn bitmap_offset<T: Into<usize> + Copy>(size: T) -> (usize, u8) {
+pub fn bitmap_offset<T: Into<u32> + Copy>(size: T) -> (usize, u8) {
     let index = size.into();
-    (index >> 3, 1 << (7 - (index & 7) as u8))
+    ((index >> 3) as usize, 1 << (7 - (index & 7) as u8))
 }
 
 /// 解码二进制数据
